refactor/app: Add default constructor and use Status enum
- Add default constructor for Result
- Use Status enum across classes
